requirements.txt
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
cryptography==3.0
decorator==4.4.2
idna==2.10
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycparser==2.20
pylint==2.6.0
python-telegram-bot==12.8
requests==2.24.0
six==1.15.0
toml==0.10.1
tornado==6.0.4
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
